// frontend/src/app/modules/admin/interfaces/ICertification.ts
export interface ICertification {
    userCertificationGuid: string;
    certificationGuid: string;
    companyGuid: string;
    companyName: string;
    skillGuid: string;
    skillTitle: string;
    certificationTitle: string;
    certificationIssueDate: string;
    certificationExpirationDate: string;
    certificationCredentialId: string;
    certificationCredentialUrl: string;
    certificationPhotoName: string;
    certificationDescription: string;
}